fix(simulator,evals): retain command diagnostics, dedupe effect, split grading and phoenix
Runtime acquisition failures projected only an exit code. execEffect now
routes through the existing capture path and retains BOTH streams, because
build tools split diagnostics across them: npm summarizes the lifecycle
failure on stderr while the compiler it invoked reports the real errors on
stdout, so keeping only the non-empty stream still hid the cause. The
staging directory is removed during cleanup, which had made that cause
unrecoverable from a durable evaluation result.

Published packages pinned `effect` to an exact version. A consumer whose
own tree resolves any newer patch therefore installs nested private copies,
and Effect's nominal types make duplicate instances a compile error rather
than merely wasteful. Declaring a range lets a consumer supply one instance.
A NanoClaw workspace build that previously failed with ten instance-identity
errors now compiles against a single deduped effect.

Two runtime tests parked forever instead of failing. A runtime arms its
startup deadline several fiber hops after its driver hands back a session,
so a fixture that had observed acquisition had not yet observed the
deadline; TestClock.adjust wakes only sleepers already registered and
anchors a later registration to the clock it has already advanced.
expireStartupDeadline waits for the deadline to appear among the scheduled
wake-ups, keeping registration order out of the outcome, and dies rather
than parking when no deadline is armed. The evaluation store test observes
real time because its SQLite retry paths sleep on the live clock.

grading.ts and phoenix.ts each suppressed size and complexity rules for
their whole file. Splitting them into cohesive modules removes every
file-wide suppression and scopes the Phoenix SDK's promise-type exception to
the adapter that needs it. Both keep their exact public surface minus
re-exports that no consumer imported.

Closes #922
Closes #925
