+++ packages/evals/package.json
@@ -26,7 +26,7 @@
     "@effect/sql-sqlite-node": "^0.53.0",
     "@moltzap/protocol": "workspace:*",
     "@moltzap/simulator": "workspace:*",
-    "effect": "3.22.0"
+    "effect": "^3.22.0"
   },
   "devDependencies": {
     "@effect/vitest": "^0.30.0",
