+++ packages/evals/src/phoenix.ts
@@ -1,63 +1,61 @@
-/* eslint-disable max-lines -- Phoenix protocol adaptation stays isolated from local report authority. */
 /** @file Idempotent Phoenix materialization for completed evaluation reports. */
 
-import {
-  createClient,
-  HttpError,
-  type PhoenixClient,
-  type Types,
-} from "@arizeai/phoenix-client";
-import { createDataset, getDataset } from "@arizeai/phoenix-client/datasets";
-import {
-  createExperiment,
-  getExperimentRuns,
-} from "@arizeai/phoenix-client/experiments";
+import { createClient, type PhoenixClient } from "@arizeai/phoenix-client";
 import { getExperimentUrl } from "@arizeai/phoenix-client/utils/urlUtils";
-import {
-  jsonValue,
-  type JsonValue as JsonValueType,
-} from "@moltzap/simulator/ledger";
 import {
   Config,
   Context,
-  DateTime,
   Effect,
   Layer,
   Option,
   Redacted,
   Schema,
 } from "effect";
-import type { CriterionAssessment } from "./grading.js";
 import { conditionId } from "./model.js";
 import {
-  canonicalJson,
+  describeUnknown,
+  type PhoenixRequestFailed,
+  type PhoenixRun,
+} from "./phoenix-client.js";
+import { datasetExampleId, ensureDataset } from "./phoenix-dataset.js";
+import {
+  phoenixAttemptEvaluations,
+  upsertEvaluation,
+} from "./phoenix-evaluation.js";
+import { ensureExperiment } from "./phoenix-experiment.js";
+import {
+  type PhoenixPublicationError,
+  type PublicationContext,
+  encodingError,
+} from "./phoenix-publication.js";
+import { ensureRun, fetchRuns } from "./phoenix-run.js";
+import {
   type CompletedEvaluationReport,
   digestEvaluationReport,
   evaluationReportDigest,
   type EvaluationConditionPlan,
-  type EvaluationReportPlan,
-  JudgePolicySnapshot,
-  terminalAttempt,
-  type EvaluationReportDigest,
   type EvaluationReportValidationError,
-  type EvidenceRejectedAttempt,
-  type JudgingUnavailableAttempt,
   type TerminalAttempt,
   validateCompletedEvaluationReport,
 } from "./sweep.js";
 
-const DATASET_NAME = "moltzap-evaluations";
-const DATASET_DESCRIPTION =
-  "MoltZap code-first behavioral evaluation cases (schema v1).";
-const PHOENIX_PUBLICATION_FORMAT_VERSION = 1;
-const FIRST_REPETITION = 1;
-const PAGE_SIZE = 100;
-
-type PhoenixDataset = Awaited<ReturnType<typeof getDataset>>;
-type PhoenixExperiment = Types["V1"]["components"]["schemas"]["Experiment"];
-type PhoenixExperimentsPage =
-  Types["V1"]["components"]["schemas"]["ListExperimentsResponseBody"];
-type PhoenixRun = Awaited<ReturnType<typeof getExperimentRuns>>["runs"][number];
+/** Re-exports the public API from `./phoenix-publication.js`. */
+export { PhoenixPublicationConflict } from "./phoenix-publication.js";
+/** Re-exports the public API from `./phoenix-dataset.js`. */
+export {
+  type PhoenixDatasetCatalog,
+  findPhoenixDataset,
+  phoenixCatalogExamples,
+  reconcilePhoenixDatasetCatalog,
+} from "./phoenix-dataset.js";
+/** Re-exports the public API from `./phoenix-experiment.js`. */
+export {
+  type PhoenixExperimentDatasetReference,
+  phoenixExperimentProvenance,
+  phoenixPublishedDatasetVersion,
+} from "./phoenix-experiment.js";
+/** Re-exports the public API from `./phoenix-evaluation.js`. */
+export { phoenixAttemptEvaluations } from "./phoenix-evaluation.js";
 
 /** Published experiment location for one runtime condition. */
 class PhoenixExperimentPublication extends Schema.Class<PhoenixExperimentPublication>(
@@ -77,39 +75,6 @@ class PhoenixPublication extends Schema.Class<PhoenixPublication>(
   experiments: Schema.NonEmptyArray(PhoenixExperimentPublication),
 }) {}
 
-/** Phoenix or its transport rejected one request. */
-class PhoenixRequestFailed extends Schema.TaggedError<PhoenixRequestFailed>()(
-  "PhoenixRequestFailed",
-  {
-    operation: Schema.NonEmptyString,
-    detail: Schema.NonEmptyString,
-    status: Schema.optional(Schema.Int),
-  },
-) {}
-
-/** A stable publication identity already names different remote state. */
-export class PhoenixPublicationConflict extends Schema.TaggedError<PhoenixPublicationConflict>()(
-  "PhoenixPublicationConflict",
-  {
-    resource: Schema.Literal("dataset", "experiment", "run"),
-    identity: Schema.NonEmptyString,
-    detail: Schema.NonEmptyString,
-  },
-) {}
-
-/** Validated local data could not be represented by the Phoenix API. */
-export class PhoenixPublicationEncodingError extends Schema.TaggedError<PhoenixPublicationEncodingError>()(
-  "PhoenixPublicationEncodingError",
-  {
-    detail: Schema.NonEmptyString,
-  },
-) {}
-
-type PhoenixPublicationError =
-  | PhoenixRequestFailed
-  | PhoenixPublicationConflict
-  | PhoenixPublicationEncodingError;
-
 type PublishFailure = EvaluationReportValidationError | PhoenixPublicationError;
 
 /** Publishes immutable completed reports as a Phoenix comparison view. */
@@ -123,948 +88,6 @@ export interface PhoenixPublisherService {
 export class PhoenixPublisher extends Context.Tag(
   "@moltzap/evals/PhoenixPublisher",
 )<PhoenixPublisher, PhoenixPublisherService>() {}
-
-function describeUnknown(cause: unknown): string {
-  if (cause instanceof Error && cause.message.length > 0) {
-    return cause.message;
-  }
-  const detail = String(cause);
-  return detail.length > 0 ? detail : "unknown Phoenix failure";
-}
-
-function requestFailure(
-  operation: string,
-  cause: unknown,
-): PhoenixRequestFailed {
-  return PhoenixRequestFailed.make({
-    operation,
-    detail: describeUnknown(cause),
-    ...(cause instanceof HttpError ? { status: cause.status } : {}),
-  });
-}
-
-/** The only Promise-to-Effect adaptation used by the publisher. */
-/* eslint-disable agent-code-guard/promise-type -- the Phoenix SDK exposes Promise APIs, which enter Effect only through this adapter. */
-function phoenixRequest<A>(
-  operation: string,
-  evaluate: () => Promise<A>,
-): Effect.Effect<A, PhoenixRequestFailed> {
-  return Effect.tryPromise({
-    try: evaluate,
-    catch: (cause) => requestFailure(operation, cause),
-  });
-}
-/* eslint-enable agent-code-guard/promise-type -- the Phoenix SDK boundary ends here. */
-
-function encodingError(detail: string): PhoenixPublicationEncodingError {
-  return PhoenixPublicationEncodingError.make({ detail });
-}
-
-function conflict(
-  resource: PhoenixPublicationConflict["resource"],
-  identity: string,
-  detail: string,
-): PhoenixPublicationConflict {
-  return PhoenixPublicationConflict.make({ resource, identity, detail });
-}
-
-function canonicalUnknown(
-  value: unknown,
-): Effect.Effect<string, PhoenixPublicationEncodingError> {
-  return Schema.decodeUnknown(jsonValue)(value).pipe(
-    Effect.map(canonicalJson),
-    Effect.mapError((cause) =>
-      encodingError(`Phoenix value is not JSON: ${cause.message}`),
-    ),
-  );
-}
-
-function sameJson(
-  left: unknown,
-  right: unknown,
-): Effect.Effect<boolean, PhoenixPublicationEncodingError> {
-  return Effect.all({
-    left: canonicalUnknown(left),
-    right: canonicalUnknown(right),
-  }).pipe(Effect.map(({ left, right }) => left === right));
-}
-
-function sorted(values: readonly string[]): readonly string[] {
-  return [...values].sort((left, right) => left.localeCompare(right));
-}
-
-/**
- * Canonical Phoenix dataset rows derived from the immutable case catalog.
- * @param plan Immutable case and condition catalog.
- * @returns Stable Phoenix examples in catalog order.
- */
-export function phoenixCatalogExamples(plan: EvaluationReportPlan) {
-  return plan.cases.map((casePlan) => {
-    const slices = sorted(casePlan.slices);
-    return {
-      id: casePlan.id,
-      input: {
-        caseId: casePlan.id,
-        definitionId: casePlan.definitionId,
-        name: casePlan.name,
-        description: casePlan.description,
-        rubric: casePlan.rubric,
-        criterionIds: [...casePlan.criterionIds],
-      },
-      output: {},
-      metadata: { slices },
-    };
-  });
-}
-
-/** SDK-readable dataset fields used during remote reconciliation. */
-export interface PhoenixDatasetCatalog {
-  readonly name: string;
-  readonly description?: string | null;
-  readonly examples: ReadonlyArray<{
-    readonly id: string;
-    readonly input: unknown;
-    readonly output?: unknown;
-    readonly metadata?: unknown;
-  }>;
-}
-
-function datasetExampleProjection(
-  example: PhoenixDatasetCatalog["examples"][number],
-) {
-  return {
-    id: example.id,
-    input: example.input,
-    output: example.output ?? {},
-    metadata: example.metadata ?? {},
-  };
-}
-
-/**
- * Reconcile remote dataset identity and examples without requiring a live
- * Phoenix client.
- * @param dataset Remote dataset projection.
- * @param plan Immutable local report plan.
- * @returns Completion when the projections agree.
- */
-export function reconcilePhoenixDatasetCatalog(
-  dataset: PhoenixDatasetCatalog,
-  plan: EvaluationReportPlan,
-): Effect.Effect<
-  void,
-  PhoenixPublicationConflict | PhoenixPublicationEncodingError
-> {
-  return Effect.gen(function* () {
-    if (dataset.name !== DATASET_NAME) {
-      return yield* Effect.fail(
-        conflict(
-          "dataset",
-          DATASET_NAME,
-          `remote dataset is named ${dataset.name}`,
-        ),
-      );
-    }
-    if (dataset.description !== DATASET_DESCRIPTION) {
-      return yield* Effect.fail(
-        conflict("dataset", DATASET_NAME, "remote dataset description differs"),
-      );
-    }
-    const actual = dataset.examples
-      .map(datasetExampleProjection)
-      .sort((left, right) => left.id.localeCompare(right.id));
-    const expected = phoenixCatalogExamples(plan).sort((left, right) =>
-      left.id.localeCompare(right.id),
-    );
-    if (!(yield* sameJson(actual, expected))) {
-      return yield* Effect.fail(
-        conflict(
-          "dataset",
-          DATASET_NAME,
-          "remote examples differ from the report case catalog",
-        ),
-      );
-    }
-  }).pipe(Effect.withSpan("evals.reconcilePhoenixDatasetCatalog"));
-}
-
-/**
- * Find the stable dataset without asking the SDK to interpret an empty list.
- * @param client Configured Phoenix SDK client.
- * @returns The unique evaluation dataset when it exists.
- */
-export function findPhoenixDataset(
-  client: PhoenixClient,
-): Effect.Effect<
-  PhoenixDataset | undefined,
-  PhoenixRequestFailed | PhoenixPublicationConflict
-> {
-  const operation = "find evaluation dataset";
-  return Effect.gen(function* () {
-    const response = yield* phoenixRequest(operation, () =>
-      client.GET("/v1/datasets", {
-        params: { query: { name: DATASET_NAME, limit: 2, cursor: null } },
-      }),
-    );
-    const datasets = response.data?.data;
-    if (datasets === undefined) {
-      return yield* Effect.fail(
-        requestFailure(operation, "Phoenix returned no dataset page"),
-      );
-    }
-    const [dataset, ...duplicates] = datasets.filter(
-      (candidate) => candidate.name === DATASET_NAME,
-    );
-    if (duplicates.length > 0) {
-      return yield* Effect.fail(
-        conflict("dataset", DATASET_NAME, "remote identity is not unique"),
-      );
-    }
-    if (dataset === undefined) {
-      return undefined;
-    }
-    return yield* phoenixRequest("get evaluation dataset", () =>
-      getDataset({ client, dataset: { datasetId: dataset.id } }),
-    );
-  }).pipe(Effect.withSpan("evals.findPhoenixDataset"));
-}
-
-function updateDatasetCatalog(
-  client: PhoenixClient,
-  plan: EvaluationReportPlan,
-): Effect.Effect<string, PhoenixRequestFailed> {
-  return phoenixRequest("update evaluation dataset", () =>
-    createDataset({
-      client,
-      name: DATASET_NAME,
-      description: DATASET_DESCRIPTION,
-      examples: phoenixCatalogExamples(plan),
-    }),
-  ).pipe(Effect.map(({ datasetId }) => datasetId));
-}
-
-type DatasetFailure =
-  | PhoenixRequestFailed
-  | PhoenixPublicationConflict
-  | PhoenixPublicationEncodingError;
-
-function fetchExperimentsPage(
-  client: PhoenixClient,
-  datasetId: string,
-  cursor: string | null,
-): Effect.Effect<PhoenixExperimentsPage, PhoenixRequestFailed> {
-  const operation = "list evaluation experiments";
-  return phoenixRequest(operation, () =>
-    client.GET("/v1/datasets/{dataset_id}/experiments", {
-      params: {
-        path: { dataset_id: datasetId },
-        query: { cursor, limit: PAGE_SIZE },
-      },
-    }),
-  ).pipe(
-    Effect.flatMap((response) => {
-      const page: PhoenixExperimentsPage | undefined = response.data;
-      return page === undefined
-        ? Effect.fail(
-            requestFailure(operation, "Phoenix returned no experiment page"),
-          )
-        : Effect.succeed(page);
-    }),
-  );
-}
-
-function collectExperimentPage(
-  client: PhoenixClient,
-  datasetId: string,
-  page: PhoenixExperimentsPage,
-  collected: readonly PhoenixExperiment[],
-): Effect.Effect<readonly PhoenixExperiment[], PhoenixRequestFailed> {
-  const next = [...collected, ...page.data];
-  if (page.next_cursor === null) {
-    return Effect.succeed(next);
-  }
-  return fetchExperimentsPage(client, datasetId, page.next_cursor).pipe(
-    Effect.flatMap((nextPage) =>
-      collectExperimentPage(client, datasetId, nextPage, next),
-    ),
-  );
-}
-
-function listExperiments(
-  client: PhoenixClient,
-  datasetId: string,
-): Effect.Effect<readonly PhoenixExperiment[], PhoenixRequestFailed> {
-  return fetchExperimentsPage(client, datasetId, null).pipe(
-    Effect.flatMap((page) =>
-      collectExperimentPage(client, datasetId, page, []),
-    ),
-  );
-}
-
-function experimentName(
-  digest: EvaluationReportDigest,
-  condition: EvaluationConditionPlan,
-): string {
-  return `moltzap/${digest}/${condition.id}`;
-}
-
-function selectCanonicalExperiment<Experiment extends { readonly id: string }>(
-  experiments: readonly Experiment[],
-): Experiment | undefined {
-  return [...experiments].sort((left, right) =>
-    left.id.localeCompare(right.id),
-  )[0];
-}
-
-/** Dataset version fields retained by an existing Phoenix experiment. */
-export type PhoenixExperimentDatasetReference = Pick<
-  PhoenixExperiment,
-  "name" | "dataset_version_id"
->;
-
-/**
- * Find the dataset version already pinned by a report's experiments.
- * @param digest Stable local report digest.
- * @param plan Immutable local report plan.
- * @param experiments Remote experiment references.
- * @returns The one pinned version, if the report is already published.
- */
-export function phoenixPublishedDatasetVersion(
-  digest: EvaluationReportDigest,
-  plan: EvaluationReportPlan,
-  experiments: readonly PhoenixExperimentDatasetReference[],
-): Effect.Effect<Option.Option<string>, PhoenixPublicationConflict> {
-  const identities = new Set(
-    plan.conditions.map((condition) => experimentName(digest, condition)),
-  );
-  const versions = new Set(
-    experiments
-      .filter((experiment) => identities.has(experiment.name))
-      .map((experiment) => experiment.dataset_version_id),
-  );
-  const mismatch = conflict(
-    "experiment",
-    digest,
-    "report experiments use different dataset versions",
-  );
-  return versions.size > 1
-    ? Effect.fail(mismatch)
-    : Effect.succeed(Option.fromIterable(versions));
-}
-
-function ensureDataset(
-  client: PhoenixClient,
-  plan: EvaluationReportPlan,
-  digest: EvaluationReportDigest,
-): Effect.Effect<PhoenixDataset, DatasetFailure> {
-  return Effect.gen(function* () {
-    const latest = yield* findPhoenixDataset(client);
-    if (latest !== undefined) {
-      const experiments = yield* listExperiments(client, latest.id);
-      const publishedVersion = yield* phoenixPublishedDatasetVersion(
-        digest,
-        plan,
-        experiments,
-      );
-      if (Option.isSome(publishedVersion)) {
-        const published = yield* phoenixRequest(
-          "get evaluation dataset version",
-          () =>
-            getDataset({
-              client,
-              dataset: {
-                datasetId: latest.id,
-                versionId: publishedVersion.value,
-              },
-            }),
-        );
-        return yield* reconcilePhoenixDatasetCatalog(published, plan).pipe(
-          Effect.as(published),
-        );
-      }
-    }
-    yield* updateDatasetCatalog(client, plan);
-    const updated = yield* findPhoenixDataset(client);
-    if (updated !== undefined) {
-      return yield* reconcilePhoenixDatasetCatalog(updated, plan).pipe(
-        Effect.as(updated),
-      );
-    }
-    return yield* Effect.fail(
-      requestFailure(
-        "get updated evaluation dataset",
-        "dataset disappeared after update",
-      ),
-    );
-  });
-}
-
-/**
- * Runtime and judge inputs exposed on each condition experiment.
- * @param plan Immutable local report plan.
- * @param condition Runtime condition projected by the experiment.
- * @returns JSON-safe runtime and judge provenance.
- */
-export function phoenixExperimentProvenance(
-  plan: EvaluationReportPlan,
-  condition: EvaluationConditionPlan,
-): Effect.Effect<
-  Record<string, JsonValueType>,
-  PhoenixPublicationEncodingError
-> {
-  return Schema.encode(JudgePolicySnapshot)(plan.judgePolicy, {
-    onExcessProperty: "error",
-  }).pipe(
-    Effect.flatMap(Schema.decodeUnknown(jsonValue)),
-    Effect.map((judgePolicy) => ({
-      runtimeConfiguration: condition.runtimeConfiguration,
-      judgePolicy,
-    })),
-    Effect.mapError((cause) =>
-      encodingError(`cannot encode experiment provenance: ${cause.message}`),
-    ),
-  );
-}
-
-function experimentMetadata(
-  report: CompletedEvaluationReport,
-  digest: EvaluationReportDigest,
-  condition: EvaluationConditionPlan,
-): Effect.Effect<
-  Record<string, JsonValueType>,
-  PhoenixPublicationEncodingError
-> {
-  return phoenixExperimentProvenance(report.plan, condition).pipe(
-    Effect.map((provenance) => ({
-      publisher: "@moltzap/evals",
-      publicationFormatVersion: PHOENIX_PUBLICATION_FORMAT_VERSION,
-      reportId: report.reportId,
-      reportDigest: digest,
-      planDigest: report.planDigest,
-      conditionId: condition.id,
-      runtimeName: condition.runtimeName,
-      sourceRevision: report.plan.sourceRevision,
-      ...provenance,
-    })),
-  );
-}
-
-interface PublicationContext {
-  readonly client: PhoenixClient;
-  readonly dataset: PhoenixDataset;
-  readonly report: CompletedEvaluationReport;
-  readonly digest: EvaluationReportDigest;
-}
-
-function validateExperiment(
-  experiment: PhoenixExperiment,
-  context: PublicationContext,
-  condition: EvaluationConditionPlan,
-): Effect.Effect<
-  PhoenixExperiment,
-  PhoenixPublicationConflict | PhoenixPublicationEncodingError
-> {
-  const identity = experimentName(context.digest, condition);
-  return Effect.gen(function* () {
-    const expectedMetadata = yield* experimentMetadata(
-      context.report,
-      context.digest,
-      condition,
-    );
-    const metadataMatches = yield* sameJson(
-      experiment.metadata,
-      expectedMetadata,
-    );
-    const matches = [
-      experiment.name === identity,
-      experiment.description ===
-        `MoltZap evaluation report ${context.report.reportId}, condition ${condition.id}.`,
-      experiment.dataset_id === context.dataset.id,
-      experiment.dataset_version_id === context.dataset.versionId,
-      experiment.repetitions === FIRST_REPETITION,
-      metadataMatches,
-    ].every(Boolean);
-    if (!matches) {
-      return yield* Effect.fail(
-        conflict(
-          "experiment",
-          identity,
-          "remote experiment identity or metadata differs",
-        ),
-      );
-    }
-    return experiment;
-  });
-}
-
-function fetchExperiment(
-  client: PhoenixClient,
-  experimentId: string,
-): Effect.Effect<PhoenixExperiment, PhoenixRequestFailed> {
-  const operation = "get evaluation experiment";
-  return phoenixRequest(operation, () =>
-    client.GET("/v1/experiments/{experiment_id}", {
-      params: { path: { experiment_id: experimentId } },
-    }),
-  ).pipe(
-    Effect.flatMap((response) => {
-      const experiment = response.data?.data;
-      return experiment === undefined
-        ? Effect.fail(
-            requestFailure(operation, "Phoenix returned no experiment"),
-          )
-        : Effect.succeed(experiment);
-    }),
-  );
-}
-
-function reconcileExperiments(
-  experiments: readonly PhoenixExperiment[],
-  context: PublicationContext,
-  condition: EvaluationConditionPlan,
-): Effect.Effect<
-  PhoenixExperiment,
-  PhoenixPublicationConflict | PhoenixPublicationEncodingError
-> {
-  return Effect.gen(function* () {
-    const validated = yield* Effect.forEach(
-      experiments,
-      (experiment) => validateExperiment(experiment, context, condition),
-      { concurrency: 1 },
-    );
-    const canonical = selectCanonicalExperiment(validated);
-    if (canonical !== undefined) {
-      return canonical;
-    }
-    return yield* Effect.fail(
-      conflict(
-        "experiment",
-        experimentName(context.digest, condition),
-        "remote identity disappeared during reconciliation",
-      ),
-    );
-  });
-}
-
-function ensureExperiment(
-  context: PublicationContext,
-  condition: EvaluationConditionPlan,
-): Effect.Effect<PhoenixExperiment, DatasetFailure> {
-  return Effect.gen(function* () {
-    const identity = experimentName(context.digest, condition);
-    const experiments = yield* listExperiments(
-      context.client,
-      context.dataset.id,
-    );
-    const matches = experiments.filter(
-      (experiment) => experiment.name === identity,
-    );
-    if (matches.length > 0) {
-      return yield* reconcileExperiments(matches, context, condition);
-    }
-    const metadata = yield* experimentMetadata(
-      context.report,
-      context.digest,
-      condition,
-    );
-    const created = yield* phoenixRequest("create evaluation experiment", () =>
-      createExperiment({
-        client: context.client,
-        datasetId: context.dataset.id,
-        datasetVersionId: context.dataset.versionId,
-        experimentName: identity,
-        experimentDescription: `MoltZap evaluation report ${context.report.reportId}, condition ${condition.id}.`,
-        experimentMetadata: metadata,
-        repetitions: FIRST_REPETITION,
-      }),
-    );
-    const remote = yield* fetchExperiment(context.client, created.id);
-    const refreshed = yield* listExperiments(
-      context.client,
-      context.dataset.id,
-    );
-    const candidates = new Map(
-      refreshed
-        .filter((experiment) => experiment.name === identity)
-        .map((experiment) => [experiment.id, experiment] as const),
-    );
-    candidates.set(remote.id, remote);
-    return yield* reconcileExperiments(
-      Array.from(candidates.values()),
-      context,
-      condition,
-    );
-  });
-}
-
-function encodeAttempt(
-  attempt: TerminalAttempt,
-): Effect.Effect<JsonValueType, PhoenixPublicationEncodingError> {
-  return Schema.encode(terminalAttempt)(attempt, {
-    onExcessProperty: "error",
-  }).pipe(
-    Effect.flatMap(Schema.decodeUnknown(jsonValue)),
-    Effect.mapError((cause) =>
-      encodingError(
-        `cannot encode attempt ${attempt.attemptId}: ${cause.message}`,
-      ),
-    ),
-  );
-}
-
-function runError(attempt: TerminalAttempt): string | null {
-  switch (attempt._tag) {
-    case "RunFailedAttempt":
-      return `${attempt._tag}: ${attempt.detail}`;
-    case "LedgerAllocationFailedAttempt":
-      return `${attempt._tag}: ${attempt.failure.detail}`;
-    case "AssessedAttempt":
-    case "EvidenceRejectedAttempt":
-    case "JudgingUnavailableAttempt":
-      return null;
-    default:
-      return attempt;
-  }
-}
-
-interface ExpectedRun {
-  readonly datasetExampleId: string;
-  readonly output: JsonValueType;
-  readonly error: string | null;
-  readonly startTime: string;
-  readonly endTime: string;
-}
-
-function expectedRun(
-  attempt: TerminalAttempt,
-  datasetExampleId: string,
-): Effect.Effect<ExpectedRun, PhoenixPublicationEncodingError> {
-  return encodeAttempt(attempt).pipe(
-    Effect.map((output) => ({
-      datasetExampleId,
-      output,
-      error: runError(attempt),
-      startTime: DateTime.formatIso(attempt.startedAt),
-      endTime: DateTime.formatIso(attempt.completedAt),
-    })),
-  );
-}
-
-function validateRun(
-  remote: PhoenixRun,
-  expected: ExpectedRun,
-  attemptId: string,
-): Effect.Effect<
-  PhoenixRun,
-  PhoenixPublicationConflict | PhoenixPublicationEncodingError
-> {
-  return Effect.gen(function* () {
-    const outputMatches = yield* sameJson(remote.output, expected.output);
-    const matches = [
-      remote.datasetExampleId === expected.datasetExampleId,
-      remote.error === expected.error,
-      remote.startTime.toISOString() === expected.startTime,
-      remote.endTime.toISOString() === expected.endTime,
-      outputMatches,
-    ].every(Boolean);
-    if (!matches) {
-      return yield* Effect.fail(
-        conflict("run", attemptId, "remote experiment run differs"),
-      );
-    }
-    return remote;
-  });
-}
-
-function createRun(
-  client: PhoenixClient,
-  experimentId: string,
-  expected: ExpectedRun,
-): Effect.Effect<string, PhoenixRequestFailed> {
-  const operation = "create evaluation experiment run";
-  return phoenixRequest(operation, () =>
-    client.POST("/v1/experiments/{experiment_id}/runs", {
-      params: { path: { experiment_id: experimentId } },
-      body: {
-        dataset_example_id: expected.datasetExampleId,
-        output: expected.output,
-        repetition_number: FIRST_REPETITION,
-        start_time: expected.startTime,
-        end_time: expected.endTime,
-        error: expected.error,
-      },
-    }),
-  ).pipe(
-    Effect.flatMap((response) => {
-      const run = response.data?.data;
-      return run === undefined
-        ? Effect.fail(
-            requestFailure(
-              operation,
-              "Phoenix returned no experiment-run identity",
-            ),
-          )
-        : Effect.succeed(run.id);
-    }),
-  );
-}
-
-function fetchRuns(
-  client: PhoenixClient,
-  experimentId: string,
-): Effect.Effect<readonly PhoenixRun[], PhoenixRequestFailed> {
-  return phoenixRequest("get evaluation experiment runs", () =>
-    getExperimentRuns({ client, experimentId, pageSize: PAGE_SIZE }),
-  ).pipe(Effect.map(({ runs }) => runs));
-}
-
-function uniqueRunByExample(
-  runs: readonly PhoenixRun[],
-  datasetExampleId: string,
-  attemptId: string,
-): Effect.Effect<PhoenixRun | undefined, PhoenixPublicationConflict> {
-  const matching = runs.filter(
-    (run) => run.datasetExampleId === datasetExampleId,
-  );
-  const [run, ...duplicates] = matching;
-  return duplicates.length > 0
-    ? Effect.fail(
-        conflict("run", attemptId, "remote run identity is not unique"),
-      )
-    : Effect.succeed(run);
-}
-
-interface RunContext {
-  readonly client: PhoenixClient;
-  readonly experimentId: string;
-  readonly runs: readonly PhoenixRun[];
-}
-
-function recoverConcurrentRun(
-  context: RunContext,
-  expected: ExpectedRun,
-  attemptId: string,
-  creationError: PhoenixRequestFailed,
-): Effect.Effect<string, DatasetFailure> {
-  return Effect.gen(function* () {
-    const refreshed = yield* fetchRuns(context.client, context.experimentId);
-    const raced = yield* uniqueRunByExample(
-      refreshed,
-      expected.datasetExampleId,
-      attemptId,
-    );
-    if (raced === undefined) {
-      return yield* Effect.fail(creationError);
-    }
-    const validated = yield* validateRun(raced, expected, attemptId);
-    return validated.id;
-  });
-}
-
-function createOrRecoverRun(
-  context: RunContext,
-  expected: ExpectedRun,
-  attemptId: string,
-): Effect.Effect<string, DatasetFailure> {
-  return createRun(context.client, context.experimentId, expected).pipe(
-    Effect.catchTag("PhoenixRequestFailed", (error) =>
-      error.status === 409
-        ? recoverConcurrentRun(context, expected, attemptId, error)
-        : Effect.fail(error),
-    ),
-  );
-}
-
-function ensureRun(
-  context: RunContext,
-  attempt: TerminalAttempt,
-  datasetExampleId: string,
-): Effect.Effect<string, DatasetFailure> {
-  return Effect.gen(function* () {
-    const expected = yield* expectedRun(attempt, datasetExampleId);
-    const existing = yield* uniqueRunByExample(
-      context.runs,
-      datasetExampleId,
-      attempt.attemptId,
-    );
-    if (existing !== undefined) {
-      const validated = yield* validateRun(
-        existing,
-        expected,
-        attempt.attemptId,
-      );
-      return validated.id;
-    }
-    return yield* createOrRecoverRun(context, expected, attempt.attemptId);
-  });
-}
-
-function score(verdict: CriterionAssessment["verdict"]): number | null {
-  switch (verdict) {
-    case "passed":
-      return 1;
-    case "failed":
-      return 0;
-    case "undecided":
-      return null;
-    default:
-      return verdict;
-  }
-}
-
-interface ExpectedEvaluation {
-  readonly name: string;
-  readonly annotatorKind: "CODE" | "LLM";
-  readonly result: {
-    readonly label: string;
-    readonly score: number | null;
-    readonly explanation: string;
-  } | null;
-  readonly error: string | null;
-  readonly metadata: Record<string, JsonValueType>;
-}
-
-function assessmentEvaluation(
-  assessment: CriterionAssessment,
-  reportDigest: EvaluationReportDigest,
-  condition: EvaluationConditionPlan,
-): ExpectedEvaluation {
-  return {
-    name: assessment.criterionId,
-    annotatorKind: assessment._tag === "CodeAssessment" ? "CODE" : "LLM",
-    result: {
-      label: assessment.verdict,
-      score: score(assessment.verdict),
-      explanation:
-        assessment._tag === "CodeAssessment"
-          ? assessment.detail
-          : assessment.rationale,
-    },
-    error: null,
-    metadata: {
-      source: assessment._tag === "CodeAssessment" ? "code" : "model",
-      criterionId: assessment.criterionId,
-      citations: [...assessment.citations],
-      reportDigest,
-      conditionId: condition.id,
-    },
-  };
-}
-
-function evidenceErrorEvaluation(
-  attempt: EvidenceRejectedAttempt,
-  reportDigest: EvaluationReportDigest,
-  condition: EvaluationConditionPlan,
-): ExpectedEvaluation {
-  return {
-    name: "moltzap.evidence",
-    annotatorKind: "CODE",
-    result: null,
-    error: attempt.detail,
-    metadata: {
-      source: "code",
-      reportDigest,
-      conditionId: condition.id,
-    },
-  };
-}
-
-function judgeErrorEvaluations(
-  attempt: JudgingUnavailableAttempt,
-  reportDigest: EvaluationReportDigest,
-  condition: EvaluationConditionPlan,
-): readonly ExpectedEvaluation[] {
-  const error = `${attempt.error._tag}: ${attempt.error.detail}`;
-  const failures: readonly ExpectedEvaluation[] =
-    attempt.pendingCriterionIds.map(
-      (criterionId): ExpectedEvaluation => ({
-        name: criterionId,
-        annotatorKind: "LLM",
-        result: null,
-        error,
-        metadata: {
-          source: "model",
-          criterionId,
-          reportDigest,
-          conditionId: condition.id,
-        },
-      }),
-    );
-  return [
-    ...attempt.codeAssessments.map((assessment) =>
-      assessmentEvaluation(assessment, reportDigest, condition),
-    ),
-    ...failures,
-  ];
-}
-
-/**
- * Pure Phoenix assessment rows for one terminal local attempt.
- * @param attempt Validated terminal matrix attempt.
- * @param reportDigest Stable local report digest.
- * @param condition Runtime condition for the attempt.
- * @returns Assessment rows to materialize on the experiment run.
- */
-export function phoenixAttemptEvaluations(
-  attempt: TerminalAttempt,
-  reportDigest: EvaluationReportDigest,
-  condition: EvaluationConditionPlan,
-): readonly ExpectedEvaluation[] {
-  switch (attempt._tag) {
-    case "AssessedAttempt":
-      return attempt.grade.assessments.map((assessment) =>
-        assessmentEvaluation(assessment, reportDigest, condition),
-      );
-    case "EvidenceRejectedAttempt":
-      return [evidenceErrorEvaluation(attempt, reportDigest, condition)];
-    case "JudgingUnavailableAttempt":
-      return judgeErrorEvaluations(attempt, reportDigest, condition);
-    case "RunFailedAttempt":
-    case "LedgerAllocationFailedAttempt":
-      return [];
-    default:
-      return attempt;
-  }
-}
-
-function upsertEvaluation(
-  client: PhoenixClient,
-  runId: string,
-  attempt: TerminalAttempt,
-  evaluation: ExpectedEvaluation,
-): Effect.Effect<void, PhoenixRequestFailed> {
-  return phoenixRequest("upsert evaluation assessment", () =>
-    client.POST("/v1/experiment_evaluations", {
-      body: {
-        experiment_run_id: runId,
-        name: evaluation.name,
-        annotator_kind: evaluation.annotatorKind,
-        start_time: DateTime.formatIso(attempt.startedAt),
-        end_time: DateTime.formatIso(attempt.completedAt),
-        result: evaluation.result,
-        error: evaluation.error,
-        metadata: evaluation.metadata,
-      },
-    }),
-  ).pipe(Effect.asVoid);
-}
-
-function datasetExampleId(
-  dataset: PhoenixDataset,
-  caseId: string,
-): Effect.Effect<string, PhoenixPublicationConflict> {
-  const examples = dataset.examples.filter((example) => example.id === caseId);
-  const [example, ...duplicates] = examples;
-  if (example === undefined || duplicates.length > 0) {
-    return Effect.fail(
-      conflict(
-        "dataset",
-        DATASET_NAME,
-        `case ${caseId} does not have exactly one stable example`,
-      ),
-    );
-  }
-  return Effect.succeed(example.nodeId);
-}
 
 interface ConditionPublicationContext {
   readonly publication: PublicationContext;
@@ -1218,4 +241,3 @@ export const phoenixPublisherLive = Layer.effect(
     return makePhoenixPublisher(client, baseUrl);
   }).pipe(Effect.withSpan("PhoenixPublisherLive")),
 );
-/* eslint-enable max-lines -- Phoenix protocol adaptation ends here. */
