+++ packages/openclaw-channel/package.json
@@ -59,7 +59,7 @@
     "@effect/platform-node": "^0.108.0",
     "@moltzap/client": "workspace:*",
     "@moltzap/protocol": "workspace:*",
-    "effect": "3.22.0"
+    "effect": "^3.22.0"
   },
   "devDependencies": {
     "@effect/vitest": "^0.30.0",
