+++ packages/server/package.json
@@ -96,7 +96,7 @@
     "@opentelemetry/sdk-trace-base": "2.7.1",
     "@opentelemetry/sdk-trace-node": "2.7.1",
     "@opentelemetry/semantic-conventions": "1.41.1",
-    "effect": "3.22.0",
+    "effect": "^3.22.0",
     "kysely": "^0.28.2",
     "kysely-pglite": "^0.6.1",
     "pg": "^8.12.0",
