+++ packages/simulator/package.json
@@ -123,7 +123,7 @@
     "@moltzap/openclaw-channel": "workspace:^",
     "@moltzap/protocol": "workspace:^",
     "@moltzap/server-core": "workspace:*",
-    "effect": "3.22.0",
+    "effect": "^3.22.0",
     "openclaw": "2026.6.33"
   },
   "devDependencies": {
