+++ packages/simulator/src/runtime/nanoclaw/runtime.test.ts
@@ -23,10 +23,10 @@ import {
   Schema,
   Scope,
   Stream,
-  TestClock,
 } from "effect";
 import { describe } from "vitest";
 import { RuntimeAcquisitionFailed } from "../process.js";
+import { expireStartupDeadline } from "../process.test-utils.js";
 import type { InstallMode } from "../packages.js";
 import type { NanoclawGatewaySession } from "./gateway.js";
 import {
@@ -47,7 +47,10 @@ const AGENT_KEY = redactedAgentKey(AGENT_KEY_TEXT);
 const READY_OUTPUT = 'level=INFO message="MoltZap connected" channel=moltzap';
 const ROUTER_URL = serverBaseUrl("http://127.0.0.1:43123");
 const PROCESS_EXIT_CODE = 23;
-const STARTUP_TIMEOUT = Duration.seconds(17);
+// `awaitProcessReady` polls readiness on a fixed interval, and expiring this
+// budget on the test clock costs one round of real timers per poll it covers.
+// A small multiple of that interval still exercises repeated polling.
+const STARTUP_TIMEOUT = Duration.millis(500);
 const MODEL_ID = "test/model";
 const PROCESS_WAIT_FAILURE = "process wait failed";
 type ProcessWaitFailure = typeof PROCESS_WAIT_FAILURE;
@@ -328,8 +331,7 @@ function readinessFailureTest() {
         connection,
       }),
     ).pipe(Effect.flip, Effect.fork);
-    yield* Deferred.await(fixture.processInput);
-    yield* TestClock.adjust(STARTUP_TIMEOUT);
+    yield* expireStartupDeadline(STARTUP_TIMEOUT);
     const observed = yield* Fiber.join(acquiring);
 
     assert.instanceOf(observed, RuntimeAcquisitionFailed);
